Refactor to use class
This commit also fixes the length of Q. It should be as long as all
actions from 0 to T, so T+1.

The names of the output files have also been changed for clarity.

The class also includes a reset method that can be used in between runs
to clear the parameters and calculated values/policies,